fix DisableOwnerWorkspaceExec breaking other tests
